Base/Toponaming: Refactor to extract filters to new file
