+++ tests/src/Base/CMakeLists.txt
@@ -2,6 +2,7 @@ target_sources(
     Tests_run
         PRIVATE
             ${CMAKE_CURRENT_SOURCE_DIR}/Axis.cpp
+            ${CMAKE_CURRENT_SOURCE_DIR}/Base64.cpp
             ${CMAKE_CURRENT_SOURCE_DIR}/Bitmask.cpp
             ${CMAKE_CURRENT_SOURCE_DIR}/BoundBox.cpp
             ${CMAKE_CURRENT_SOURCE_DIR}/Builder3D.cpp
